fix: harden form parsing, validation, and report rendering
- parse_review: a JSON payload with an empty findings list is a valid
  clean review — return it (with its overall verdict) instead of falling
  through to the legacy parser and dropping the recommendation.
- from_dict: tolerate wrong-typed values (lists/dicts) in enum and string
  fields, defaulting instead of raising; validation DRYed into helpers.
- _build_gate: check for sentence-transformers before dispatching the
  council so a missing gate-v2 extra fails before any API spend.
- report: escape model-controlled text in table cells and inline lines;
  show '(no findings raised)' for clean reviews with an overall verdict.
- _slice_outer_json: start from the earliest bracket so a prose-wrapped
  top-level array parses.
- mypy strict clean; drop a duplicated test; fix a deprecated import.

+++ src/council_gate/clustering.py
@@ -93,7 +93,7 @@ def _cosine(a: list[float], b: list[float]) -> float:
     nb = sum(x * x for x in b) ** 0.5
     if na == 0 or nb == 0:
         return 0.0
-    return dot / (na * nb)
+    return float(dot / (na * nb))
 
 
 def _build_cluster(members: list[tuple[str, Finding]]) -> FindingCluster:
